feat(db): abstract orm layer
