+++ orm/mongoose.js
@@ -1,0 +1,5 @@
+module.exports = new Proxy({}, {
+  get(target, prop) {
+    throw new Error(`Mongoose adapter not implemented: ${prop}`);
+  }
+});
