+++ orm/postgres.js
@@ -1,0 +1,5 @@
+module.exports = new Proxy({}, {
+  get(target, prop) {
+    throw new Error(`Postgres adapter not implemented: ${prop}`);
+  }
+});
