+++ util/database.js
@@ -1,30 +1,19 @@
-const Sequelize = require("sequelize");
+const Config = require('../config/wotlwedu');
 
-const Config = require("../config/wotlwedu");
+let adapter;
 
-const options = {
-  host: Config.db_host,
-  dialect: "mariadb",
-  omitNull: false,
-};
-
-if (Config.db_logging === false) {
-  options.logging = false;
+switch ((Config.db_type || 'sequelize').toLowerCase()) {
+  case 'mongoose':
+    adapter = require('../orm/mongoose');
+    break;
+  case 'pg':
+  case 'postgres':
+  case 'postgresql':
+    adapter = require('../orm/postgres');
+    break;
+  default:
+    adapter = require('../orm/sequelize');
+    break;
 }
 
-/* TO DO: Make this configurable */
-options.pool = {
-  max: 5,
-  min: 0,
-  acquire: 30000,
-  idle: 10000
-}
-
-const sequelize = new Sequelize(
-  Config.db_database,
-  Config.db_user,
-  Config.db_password,
-  options
-);
-
-module.exports = sequelize;
+module.exports = adapter;
